+++ src/bcf/mod.rs
@@ -1230,7 +1230,6 @@ mod tests {
                 Format::VCF,
             )
             .expect("Error opening file.");
-            let header = writer.header().clone();
 
             // Setup empty record, filled below.
             let mut record = writer.empty_record();
@@ -1254,10 +1253,10 @@ mod tests {
             record.push_id(b"first_id").unwrap();
 
             assert!(record.filters().next().is_none());
-            record.set_filters(&[header.name_to_id(b"q10").unwrap()]);
-            record.push_filter(header.name_to_id(b"s50").unwrap());
-            record.remove_filter(header.name_to_id(b"q10").unwrap(), true);
-            record.push_filter(header.name_to_id(b"q10").unwrap());
+            record.set_filters(&[b"q10"]).unwrap();
+            record.push_filter(b"s50").unwrap();
+            record.remove_filter(b"q10", true).unwrap();
+            record.push_filter(b"q10").unwrap();
 
             record.set_alleles(&[b"C", b"T", b"G"]).unwrap();