+++ src/bcf/record.rs
@@ -120,7 +120,7 @@ impl<'a, T: 'a + fmt::Debug, B: Borrow<Buffer> + 'a> Deref for BufferBacked<'a, 
 }
 
 impl<'a, T: 'a + fmt::Debug + fmt::Display, B: Borrow<Buffer> + 'a> fmt::Display
-    for BufferBacked<'a, T, B>
+for BufferBacked<'a, T, B>
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt::Display::fmt(&self.value, f)
@@ -335,69 +335,159 @@ impl Record {
 
     /// Query whether the filter with the given ID has been set.
     ///
-    /// # Arguments
+    /// This method can be used to check if a record passes filtering by using either `PASS` or `.`
     ///
-    /// - `flt_id` - The filter ID to query for.
-    pub fn has_filter(&self, flt_id: Id) -> bool {
-        if *flt_id == 0 && self.inner().d.n_flt == 0 {
-            return true;
+    /// # Example
+    /// ```rust
+    /// # use rust_htslib::bcf::{Format, Header, Writer};
+    /// # use tempfile::NamedTempFile;
+    /// # let tmp = tempfile::NamedTempFile::new().unwrap();
+    /// # let path = tmp.path();
+    /// let mut header = Header::new();
+    /// header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+    /// # let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+    /// # let mut record = vcf.empty_record();
+    /// assert!(record.has_filter(b"PASS"));
+    /// assert!(record.has_filter(b"."));
+    ///
+    /// record.push_filter(b"foo").unwrap();
+    /// assert!(record.has_filter(b"foo"));
+    /// assert!(!record.has_filter(b"PASS"))
+    /// ```
+    pub fn has_filter(&self, flt_id: &[u8]) -> bool {
+        let filter_c_str = ffi::CString::new(flt_id).unwrap();
+        match unsafe { htslib::bcf_has_filter(self.header().inner, self.inner, filter_c_str.as_ptr() as *mut i8) } {
+            1 => true,
+            _ => false
         }
-        for i in 0..(self.inner().d.n_flt as isize) {
-            if unsafe { *self.inner().d.flt.offset(i) } == *flt_id as i32 {
-                return true;
-            }
-        }
-        false
     }
 
-    /// Set the given filters IDs to the FILTER column.
+    /// Set the given filter IDs to the FILTER column.
     ///
-    /// Setting an empty slice removes all filters.
+    /// Setting an empty slice removes all filters and sets `PASS`.
     ///
-    /// # Arguments
+    /// # Example
+    /// ```rust
+    /// # use rust_htslib::bcf::{Format, Header, Writer};
+    /// # use tempfile::NamedTempFile;
+    /// # let tmp = tempfile::NamedTempFile::new().unwrap();
+    /// # let path = tmp.path();
+    /// let mut header = Header::new();
+    /// header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+    /// header.push_record(br#"##FILTER=<ID=bar,Description="a horse walks into...">"#);
+    /// # let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+    /// # let mut record = vcf.empty_record();
+    /// assert!(record.has_filter(b"PASS"));
+    /// record.set_filters(&[b"foo", b"bar"]).unwrap();
+    /// assert!(record.has_filter(b"foo"));
+    /// assert!(record.has_filter(b"bar"));
+    /// assert!(!record.has_filter(b"PASS"));
+    /// record.set_filters(&[]).unwrap();
+    /// assert!(record.has_filter(b"PASS"));
+    /// assert!(!record.has_filter(b"foo"));
+    /// // 'baz' isn't in the header
+    /// assert!(record.set_filters(&[b"foo", b"baz"]).is_err())
+    /// ```
     ///
-    /// - `flt_ids` - The identifiers of the filter values to set.
-    pub fn set_filters(&mut self, flt_ids: &[Id]) {
-        let mut flt_ids: Vec<i32> = flt_ids.iter().map(|x| **x as i32).collect();
+    /// # Errors
+    /// If any of the filter IDs do not exist in the header, an [`Error::BcfUnknownID`] is returned.
+    ///
+    pub fn set_filters(&mut self, flt_ids: &[&[u8]]) -> Result<()> {
+        // let id = self.header().name_to_id(flt_id)?;
+        let mut ids: Vec<i32> = flt_ids.iter().map(|id| self.header().name_to_id(id).and_then(|id| Ok(*id as i32))).collect::<Result<Vec<i32>>>()?;
         unsafe {
             htslib::bcf_update_filter(
                 self.header().inner,
                 self.inner,
-                flt_ids.as_mut_ptr(),
-                flt_ids.len() as i32,
+                ids.as_mut_ptr(),
+                ids.len() as i32,
             );
         }
+        Ok(())
     }
 
     /// Add the given filter to the FILTER column.
     ///
-    /// If `val` corresponds to `"PASS"` then all existing filters are removed first. If other than
-    /// `"PASS"`, then existing `"PASS"` is removed.
+    /// If `flt_id` is `PASS` or `.` then all existing filters are removed first. Otherwise,
+    /// any existing `PASS` filter is removed.
     ///
-    /// # Arguments
+    /// # Example
+    /// ```rust
+    /// # use rust_htslib::bcf::{Format, Header, Writer};
+    /// # use tempfile::NamedTempFile;
+    /// # let tmp = tempfile::NamedTempFile::new().unwrap();
+    /// # let path = tmp.path();
+    /// let mut header = Header::new();
+    /// header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+    /// # let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+    /// # let mut record = vcf.empty_record();
+    /// assert!(record.has_filter(b"PASS"));
     ///
-    /// - `flt_id` - The corresponding filter ID value to add.
-    pub fn push_filter(&mut self, flt_id: Id) {
+    /// record.push_filter(b"foo").unwrap();
+    /// assert!(record.has_filter(b"foo"));
+    /// assert!(!record.has_filter(b"PASS"));
+    /// // filter must exist in the header
+    /// assert!(record.push_filter(b"bar").is_err())
+    /// ```
+    ///
+    /// # Errors
+    /// If the `flt_id` does not exist in the header, an [`Error::BcfUnknownID`] is returned.
+    ///
+    pub fn push_filter(&mut self, flt_id: &[u8]) -> Result<()> {
+        let id = self.header().name_to_id(flt_id)?;
         unsafe {
-            htslib::bcf_add_filter(self.header().inner, self.inner, *flt_id as i32);
-        }
+            htslib::bcf_add_filter(self.header().inner, self.inner, *id as i32);
+        };
+        Ok(())
     }
 
     /// Remove the given filter from the FILTER column.
     ///
     /// # Arguments
     ///
-    /// - `val` - The corresponding filter ID to remove.
-    /// - `pass_on_empty` - Set to "PASS" when removing the last value.
-    pub fn remove_filter(&mut self, flt_id: Id, pass_on_empty: bool) {
+    /// - `flt_id` - The corresponding filter ID to remove.
+    /// - `pass_on_empty` - Set to `PASS` when removing the last filter.
+    ///
+    /// # Example
+    /// ```rust
+    /// # use rust_htslib::bcf::{Format, Header, Writer};
+    /// # use tempfile::NamedTempFile;
+    /// # let tmp = tempfile::NamedTempFile::new().unwrap();
+    /// # let path = tmp.path();
+    /// let mut header = Header::new();
+    /// header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+    /// header.push_record(br#"##FILTER=<ID=bar,Description="a horse walks into...">"#);
+    /// # let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+    /// # let mut record = vcf.empty_record();
+    /// record.set_filters(&[b"foo", b"bar"]).unwrap();
+    /// assert!(record.has_filter(b"foo"));
+    /// assert!(record.has_filter(b"bar"));
+    ///
+    /// record.remove_filter(b"foo", true).unwrap();
+    /// assert!(!record.has_filter(b"foo"));
+    /// assert!(record.has_filter(b"bar"));
+    /// // 'baz' is not in the header
+    /// assert!(record.remove_filter(b"baz", true).is_err());
+    ///
+    /// record.remove_filter(b"bar", true).unwrap();
+    /// assert!(!record.has_filter(b"bar"));
+    /// assert!(record.has_filter(b"PASS"));
+    /// ```
+    ///
+    /// # Errors
+    /// If the `flt_id` does not exist in the header, an [`Error::BcfUnknownID`] is returned.
+    ///
+    pub fn remove_filter(&mut self, flt_id: &[u8], pass_on_empty: bool) -> Result<()> {
+        let id = self.header().name_to_id(flt_id)?;
         unsafe {
             htslib::bcf_remove_filter(
                 self.header().inner,
                 self.inner,
-                *flt_id as i32,
+                *id as i32,
                 pass_on_empty as i32,
-            );
-        }
+            )
+        };
+        Ok(())
     }
 
     /// Get alleles strings.
@@ -550,8 +640,8 @@ impl Record {
     }
 
     pub fn genotypes_shared_buffer<'a, B>(&self, buffer: B) -> Result<Genotypes<'a, B>>
-    where
-        B: BorrowMut<Buffer> + Borrow<Buffer> + 'a,
+        where
+            B: BorrowMut<Buffer> + Borrow<Buffer> + 'a,
     {
         Ok(Genotypes {
             encoded: self.format_shared_buffer(b"GT", buffer).integer()?,
@@ -933,7 +1023,7 @@ impl genome::AbstractLocus for Record {
                 .rid2name(self.rid().expect("rid not set"))
                 .expect("unable to find rid in header"),
         )
-        .expect("unable to interpret contig name as UTF-8")
+            .expect("unable to interpret contig name as UTF-8")
     }
 
     fn pos(&self) -> u64 {
@@ -953,8 +1043,8 @@ pub enum GenotypeAllele {
 impl GenotypeAllele {
     /// Decode given integer according to BCF standard.
     #[deprecated(
-        since = "0.36.0",
-        note = "Please use the conversion trait From<i32> for GenotypeAllele instead."
+    since = "0.36.0",
+    note = "Please use the conversion trait From<i32> for GenotypeAllele instead."
     )]
     pub fn from_encoded(encoded: i32) -> Self {
         match (encoded, encoded & 1) {
@@ -1032,8 +1122,8 @@ impl fmt::Display for Genotype {
 /// Lazy representation of genotypes, that does no computation until a particular genotype is queried.
 #[derive(Debug)]
 pub struct Genotypes<'a, B>
-where
-    B: Borrow<Buffer> + 'a,
+    where
+        B: Borrow<Buffer> + 'a,
 {
     encoded: BufferBacked<'a, Vec<&'a [i32]>, B>,
 }
@@ -1154,14 +1244,14 @@ impl<'a, 'b, B: BorrowMut<Buffer> + Borrow<Buffer> + 'b> Info<'a, B> {
                     unsafe {
                         slice::from_raw_parts(self.buffer.borrow().inner as *const u8, ret as usize)
                     }
-                    .split(|c| *c == b',')
-                    .map(|s| {
-                        // stop at zero character
-                        s.split(|c| *c == 0u8)
-                            .next()
-                            .expect("Bug: returned string should not be empty.")
-                    })
-                    .collect(),
+                        .split(|c| *c == b',')
+                        .map(|s| {
+                            // stop at zero character
+                            s.split(|c| *c == 0u8)
+                                .next()
+                                .expect("Bug: returned string should not be empty.")
+                        })
+                        .collect(),
                     self.buffer,
                 )
             })
@@ -1403,5 +1493,88 @@ mod tests {
         assert_eq!(record.rlen(), 0);
         assert_eq!(record.sample_count(), 0);
         assert_eq!(record.pos(), 0)
+    }
+
+    #[test]
+    fn test_record_has_filter_pass_is_default() {
+        let tmp = NamedTempFile::new().unwrap();
+        let path = tmp.path();
+        let header = Header::new();
+        let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+        let record = vcf.empty_record();
+
+        assert!(record.has_filter(b"PASS"));
+        assert!(record.has_filter(b"."));
+        assert!(!record.has_filter(b"foo"))
+    }
+
+    #[test]
+    fn test_record_has_filter_custom() {
+        let tmp = NamedTempFile::new().unwrap();
+        let path = tmp.path();
+        let mut header = Header::new();
+        header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+        let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+        let mut record = vcf.empty_record();
+        record.push_filter(b"foo").unwrap();
+
+        assert!(record.has_filter(b"foo"));
+        assert!(!record.has_filter(b"PASS"))
+    }
+
+    #[test]
+    fn test_record_push_filter() {
+        let tmp = NamedTempFile::new().unwrap();
+        let path = tmp.path();
+        let mut header = Header::new();
+        header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+        let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+        let mut record = vcf.empty_record();
+        assert!(record.has_filter(b"PASS"));
+        record.push_filter(b"foo").unwrap();
+        assert!(record.has_filter(b"foo"));
+        assert!(!record.has_filter(b"PASS"));
+        assert!(record.push_filter(b"bar").is_err())
+    }
+
+    #[test]
+    fn test_record_set_filters() {
+        let tmp = NamedTempFile::new().unwrap();
+        let path = tmp.path();
+        let mut header = Header::new();
+        header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+        header.push_record(br#"##FILTER=<ID=bar,Description="a horse walks into...">"#);
+        let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+        let mut record = vcf.empty_record();
+        assert!(record.has_filter(b"PASS"));
+        record.set_filters(&[b"foo", b"bar"]).unwrap();
+        assert!(record.has_filter(b"foo"));
+        assert!(record.has_filter(b"bar"));
+        assert!(!record.has_filter(b"PASS"));
+        record.set_filters(&[]).unwrap();
+        assert!(record.has_filter(b"PASS"));
+        assert!(!record.has_filter(b"foo"));
+        assert!(record.set_filters(&[b"foo", b"baz"]).is_err())
+    }
+
+    #[test]
+    fn test_record_remove_filter() {
+        let tmp = NamedTempFile::new().unwrap();
+        let path = tmp.path();
+        let mut header = Header::new();
+        header.push_record(br#"##FILTER=<ID=foo,Description="sample is a foo fighter">"#);
+        header.push_record(br#"##FILTER=<ID=bar,Description="a horse walks into...">"#);
+        let vcf = Writer::from_path(path, &header, true, Format::VCF).unwrap();
+        let mut record = vcf.empty_record();
+        record.set_filters(&[b"foo", b"bar"]).unwrap();
+        assert!(record.has_filter(b"foo"));
+        assert!(record.has_filter(b"bar"));
+        record.remove_filter(b"foo", true).unwrap();
+        assert!(!record.has_filter(b"foo"));
+        assert!(record.has_filter(b"bar"));
+        assert!(record.remove_filter(b"baz", true).is_err());
+        record.remove_filter(b"bar", true).unwrap();
+        assert!(!record.has_filter(b"bar"));
+        assert!(record.has_filter(b"PASS"));
     }
 }